refactor: remove class writer
refs: fix/transpile-stringable
