--- config/typescript-generator.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-return [
-    'output_path' => resource_path('ts'),
-
-    'page_path' => 'Pages',
-];
